+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Merged branches and closed issues
 
+- Operations - expose sigma to users in Gaussian pulse, also change nr_sigma of in waveforms from 3 to 4 as it is defined as 4 in pulse_library (!926)
 - Linter
   - Replace deprecated numpy definitions with numpy 2.0 compatible definitions. (!952)
 - Schedules - Add `long_time_trace` and `long_time_trace_with_qubit` to perform custom integrated time traces. (!878)